Move instruction filters to fingerprint.kt file
